Moved functionality to main.py.
