+++ mochi/core/__init__.py
@@ -1,1 +1,1 @@
-from .builtins import *
+from .main import main
